Separating integration tests from unit tests
