+++ api/server/server.go
@@ -27,7 +27,6 @@ import (
 
 const (
 	defaultTimeOut = 30 * time.Second
-	natsClusterID  = "test-cluster"
 	natsClientID   = "amplifier"
 )
 
@@ -134,7 +133,7 @@ func initNats(config Config) error {
 		fmt.Errorf("amplifer is unable to connect to NATS on: %s\n%v", config.NatsURL, err)
 	}
 
-	runtime.Nats, err = stan.Connect(natsClusterID, natsClientID+strconv.Itoa(rand.Int()), stan.NatsConn(nc), stan.ConnectWait(defaultTimeOut))
+	runtime.Nats, err = stan.Connect(runtime.NatsClusterID, natsClientID+strconv.Itoa(rand.Int()), stan.NatsConn(nc), stan.ConnectWait(defaultTimeOut))
 	if err != nil {
 		return fmt.Errorf("amplifer is unable to connect to NATS-Streaming on: %s\n%v", config.NatsURL, err)
 	}
