+++ tests/cli/lookup/regex-lookup.yml
