+++ tests/cli/samples/stack.yml
@@ -3,7 +3,7 @@
   args:
      - "{{call .uniq `stack1`}}"
   options:
-     - -f ../../../api/rpc/stack/test_samples/sample-04.yml
+     - -f ../../api/rpc/stack/test_samples/sample-04.yml
   expectation: stack-id
   timeout: 2000
   delay: 1000
