+++ tests/cli/samples/stats.yml
