+++ tests/data/storage/etcd/store_test.proto
