+++ tests/rpc/stats_test.go
