+++ tests/rpc/topic_test.go
